requirements.txt
numpy>=1.24
Pillow>=10.0
scipy>=1.11
matplotlib>=3.7
reportlab>=4.0
